+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changes in Matrix iOS SDK in 0.11.2 (2018-08-)
 
 Improvements:
  * MXRestClient: Add testUserRegistration to check earlier if a username can be registered.
+ * MXSession: Add MXSessionStateSyncError state and MXSession.syncError to manage homeserver resource quota on /sync requests (vector-im/riot-ios/issues/1937).
  * MXError: Add kMXErrCodeStringResourceLimitExceeded to manage homeserver resource quota (vector-im/riot-ios/issues/1937).
  * MXError: Define constant strings for keys and values that can be found in a Matrix JSON dictionary error.